+++ types/checkbox-group.d.ts
@@ -1,7 +1,8 @@
 import { ElementUIComponent, ElementUIComponentSize } from './component'
 
 /** Checkbox Group Component */
-export declare class ElCheckboxGroup extends ElementUIComponent {
+export const ElCheckboxGroup : ICheckboxGroup
+interface ICheckboxGroup extends ElementUIComponent {
   /** Size of checkbox buttons or bordered checkboxes */
   size: ElementUIComponentSize